Added Problem 23 and Fixed bug in list_factors
list_factors() was returning the sqrt of a number twice if it was a
factor. This has been fixed by making list_factors use a set instead of
a list.

Added list_proper_divisors() which returns the factors of a number, but
includes 1 and not the number itself.

Added is_abundant() which checks if a number is an abundant number.

Added problem23()